feat: 抽离b端题型枚举
