--- web/src/management/utils/constant.js
@@ -1,18 +1,0 @@
-// 问卷操作枚举
-export const QOP_MAP = {
-  ADD: 'add',
-  COPY: 'copy',
-  EDIT: 'edit'
-}
-
-export const qAbleList = ['radio', 'checkbox', 'binary-choice', 'vote']
-export const operatorOptions = [
-  {
-    label: '选择了',
-    value: 'in'
-  },
-  {
-    label: '不选择',
-    value: 'nin'
-  }
-]
